pyproject.toml
[project]
name = "dv-register-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "mysql-connector-python>=9.5.0",
    "python-telegram-bot[job-queue]>=22.5",
]
